GenerateKey now has a SendAck option.
Various other refactorings:
 * Separate key manager types from implementation.
 * Resolve golint issues in keymgr
 * Boolean pattern allowed for commands.
 * By default, new identities return acks.

+++ dbsetup.go
@@ -310,7 +310,7 @@ func createDatabases(cfg *Config) error {
 		return err
 	}
 
-	err = user.Initialize(u, kmgr, cfg.GenKeys)
+	err = user.Initialize(u, kmgr, uint32(cfg.GenKeys))
 	if err != nil {
 		return err
 	}
